Update the list of secret tokens
This updates the list of what should use a secure randomness source,
based on recent security issues.

We want to make it explicit that a session id should be securely
generated, since they are often used to provide access to a website.

Captcha strings are a kind of access token and should be generated
securely.

MFA codes are also explicitly listed.